refactor(admin): consolidate LLM provider management module
The LLM provider settings page, its state logic and styles were split across view/, composables/ and styles/. Move Llm.vue, useLlm.ts and llm.scss under admin/llm so provider management forms one module.
